requirements.txt
openai
firecrawl-py
streamlit